requirements.txt
Django~=3.2.6
africastalking~=1.2.4
jsonfield~=3.1.0
django-environ

gunicorn
django-heroku
twilio

requests~=2.26.0
waitress

djangorestframework~=3.12.4
asgiref~=3.4.1
django-cors-headers
python-dateutil
